refactor: extract SlideshowViewModel from SlideshowView
Move slideshow state, sequencer interaction, asset loading, timers and
the AVPlayer lifecycle into an ObservableObject; SlideshowView is now a
thin renderer that observes it. Logic is moved verbatim (same execution
contexts) to preserve behavior.

VERIFY ON APPLE TV before shipping: unit tests and the launch test pass,
but interactive slideshow behavior (auto-advance timers, remote left/
right/up/down, pause, video playback, focus) cannot be verified without
a device.
