+++ app/BigImmich/Slideshow/SlideshowView.swift
@@ -1,6 +1,5 @@
 import AVKit
 import ImmichAPI
-import Sentry
 import SwiftUI
 
 struct SlideshowView: View {
@@ -9,66 +8,44 @@ struct SlideshowView: View {
     let initialAssetID: AssetID?
     let onExit: (AlbumID, AlbumName, AssetID?) -> Void
 
-    @State private var slideshow: SlideshowSequencer?
-    @State private var slideshowAsset: SlideshowAsset?
-    @State private var previousAlbumID: AlbumID?
-
-    // showing details of an image (when paused)
-    @State private var userAssetIndex: Int = 0
-    @State private var userAssetsCount: Int = 0
-    @State private var userDateTime: String = ""
-    @State private var userLocation: String = ""
-
-    // current image / player in the slideshow
-    @State private var currentImage: Image?
-    @State private var currentPlayer: AVPlayer?
-    @State private var playerObserver: NSObjectProtocol?
-    @State private var timeObserverToken: Any?
-    @State private var playerIsVisible: Bool = false
-    @State private var videoPausedByUser = false
-
-    // loading assets and error reporting
-    @State private var isLoading = false
-    @State private var errors: [String] = []
-    @State private var clearErrors: DispatchWorkItem?
-    @State private var informations: [String] = []
-
-    /// loading settings
-    @State private var settings: SlideshowSettings = .init()
-
-    // slideshow + overlays
-    @State private var slideshowTimer: Timer?
-    @State private var slideshowIsRunning = true
-    @State private var showAssetDetails = false
+    @StateObject private var viewModel: SlideshowViewModel
     @State private var assetDetailsTimer: Timer?
 
-    // progress bar
-    @State private var assetProgress: Double = 0.0
-    @State private var progressBarTimer: Timer?
-
-    /// preloading assets
-    @State private var imageCache: MemoryCache<AssetID, Image>?
+    init(
+        initialAlbumID: AlbumID,
+        initialAlbumName: AlbumName,
+        initialAssetID: AssetID?,
+        onExit: @escaping (AlbumID, AlbumName, AssetID?) -> Void
+    ) {
+        self.initialAlbumID = initialAlbumID
+        self.initialAlbumName = initialAlbumName
+        self.initialAssetID = initialAssetID
+        self.onExit = onExit
+        _viewModel = StateObject(
+            wrappedValue: SlideshowViewModel(
+                initialAlbumID: initialAlbumID,
+                initialAlbumName: initialAlbumName,
+                initialAssetID: initialAssetID
+            )
+        )
+    }
 
     var body: some View {
         ZStack {
             Color.black.ignoresSafeArea()
 
             Group {
-                if isLoading {
+                if viewModel.isLoading {
                     ProgressView("Loading...")
                         .scaleEffect(1.5)
                         .frame(maxWidth: .infinity, maxHeight: .infinity)
-                } else if let player = currentPlayer {
+                } else if let player = viewModel.currentPlayer {
                     VideoPlayer(player: player)
                         .ignoresSafeArea()
                         .onAppear {
-                            playerIsVisible = true
                             player.play()
                         }
-                        .onDisappear {
-                            playerIsVisible = false
-                        }
-                } else if let image = currentImage {
+                } else if let image = viewModel.currentImage {
                     image
                         .resizable()
                         .scaledToFit()
@@ -77,639 +54,165 @@ struct SlideshowView: View {
                 }
             }
 
-            if !slideshowIsRunning {
-                VStack {
-                    Spacer()
-                    HStack {
-                        // pause icon on the bottom left
-                        Image(systemName: "pause.fill")
-                            .font(.largeTitle)
-                            .foregroundColor(.white.opacity(0.9))
-                            .padding(20)
-                            .background(
-                                Circle()
-                                    .fill(Color.black.opacity(0.4)) // semi-transparent background
-                                    .shadow(
-                                        color: Color.black.opacity(0.6),
-                                        radius: 6,
-                                        x: 0,
-                                        y: 0
-                                    ) // round shadow
-                            )
-                            .padding(.leading, 40)
-
-                        Spacer()
-
-                        // exif and stuff on the bottom right
-                        if showAssetDetails {
-                            VStack(alignment: .trailing, spacing: 4) {
-                                Text("\(userAssetIndex) / \(userAssetsCount)")
-                                    .font(.title2)
-                                    .foregroundColor(.white)
-                                    .shadow(
-                                        color: .black.opacity(0.8),
-                                        radius: 2,
-                                        x: 0,
-                                        y: 1
-                                    )
-
-                                if !(userDateTime.isEmpty) {
-                                    Text("\(userDateTime)")
-                                        .font(.headline)
-                                        .foregroundColor(.white.opacity(0.9))
-                                        .shadow(
-                                            color: .black.opacity(0.7),
-                                            radius: 1,
-                                            x: 0,
-                                            y: 1
-                                        )
-                                }
-
-                                if !(userLocation.isEmpty) {
-                                    Text("\(userLocation)")
-                                        .font(.headline)
-                                        .foregroundColor(.white.opacity(0.9))
-                                        .shadow(
-                                            color: .black.opacity(0.7),
-                                            radius: 1,
-                                            x: 0,
-                                            y: 1
-                                        )
-                                }
-                            }
-                            .padding(16)
-                            .background(Color.black.opacity(0.6))
-                            .cornerRadius(10)
-                            .onAppear {
-                                assetDetailsTimer?.invalidate()
-
-                                assetDetailsTimer = Timer.scheduledTimer(
-                                    withTimeInterval: 10,
-                                    repeats: false
-                                ) { _ in
-                                    withAnimation {
-                                        showAssetDetails = false
-                                    }
-                                }
-                            }
-                            .onDisappear {
-                                assetDetailsTimer?.invalidate()
-                                assetDetailsTimer = nil
-                            }
-                        }
-                    }
-                    .padding(.bottom, 40)
-                }
-                .transition(.opacity)
+            if !viewModel.slideshowIsRunning {
+                pausedOverlay
             }
 
-            // progress bar
-            if settings.slideshowShowProgressBar == .always {
-                GeometryReader { geometry in
-                    Rectangle()
-                        .fill(Color.gray.opacity(0.6))
-                        .frame(
-                            width: geometry.size.width * assetProgress,
-                            height: 4
-                        )
-                        .position(
-                            x: geometry.size.width / 2,
-                            y: geometry.size.height - 2
-                        )
-                }
-                .ignoresSafeArea()
+            if viewModel.settings.slideshowShowProgressBar == .always {
+                progressBar
             }
 
-            // errors and informations overlay
-            if !errors.isEmpty || !informations.isEmpty {
-                VStack {
-                    Spacer()
-                    HStack {
-                        Spacer()
-                        VStack(alignment: .trailing, spacing: 6) {
-                            ForEach(errors.indices, id: \.self) { index in
-                                Text(errors[index])
-                                    .foregroundColor(.white)
-                                    .padding(.horizontal, 12)
-                                    .padding(.vertical, 8)
-                                    .background(Color.red.opacity(0.8))
-                                    .cornerRadius(10)
-                                    .shadow(radius: 3)
-                            }
-
-                            ForEach(informations.indices, id: \.self) { index in
-                                Text(informations[index])
-                                    .foregroundColor(.white)
-                                    .padding(.horizontal, 12)
-                                    .padding(.vertical, 8)
-                                    .background(Color.blue.opacity(0.8))
-                                    .cornerRadius(10)
-                                    .shadow(radius: 3)
-                            }
-                        }
-                        .padding(.trailing, 20)
-                        .padding(.bottom, 20)
-                    }
-                }
-                .transition(.opacity.combined(with: .move(edge: .trailing)))
-                .animation(.easeInOut, value: errors)
+            if !viewModel.errors.isEmpty || !viewModel.informations.isEmpty {
+                messagesOverlay
             }
         }
         .focusable(true)
         .onAppear {
             Task {
-                imageCache = MemoryCache(countLimit: 10, megaBytesLimit: nil)
-                await initSlideshow(albumID: initialAlbumID)
+                await viewModel.start()
             }
         }
         .onDisappear {
-            stopSlideshowTimer()
-            stopProgressBarTimer()
-            stopCurrentPlayer()
+            viewModel.stop()
         }
         .onExitCommand {
-            imageCache?.clear()
+            viewModel.clearImageCache()
 
-            Task {
-                if let slideshow,
-                   let slideshowAsset = try? await slideshow.current()
-                {
-                    onExit(
-                        slideshowAsset.album.id,
-                        slideshowAsset.album.albumName,
-                        slideshowAsset.asset.id
-                    )
-                } else {
-                    onExit(initialAlbumID, initialAlbumName, initialAssetID)
-                }
+            if let asset = viewModel.slideshowAsset {
+                onExit(asset.album.id, asset.album.albumName, asset.asset.id)
+            } else {
+                onExit(initialAlbumID, initialAlbumName, initialAssetID)
             }
         }
         .onMoveCommand { direction in
             Task {
-                await handleMoveCommand(direction)
+                await viewModel.handleMoveCommand(direction)
             }
         }
         .onPlayPauseCommand {
-            togglePause()
+            viewModel.togglePause()
         }
     }
 
-    func showError(_ message: String) {
-        AppLog.shared.log(message, level: .error, source: "SlideshowView")
-        withAnimation {
-            errors.append(message)
+    private var pausedOverlay: some View {
+        VStack {
+            Spacer()
+            HStack {
+                Image(systemName: "pause.fill")
+                    .font(.largeTitle)
+                    .foregroundColor(.white.opacity(0.9))
+                    .padding(20)
+                    .background(
+                        Circle()
+                            .fill(Color.black.opacity(0.4))
+                            .shadow(
+                                color: Color.black.opacity(0.6),
+                                radius: 6,
+                                x: 0,
+                                y: 0
+                            )
+                    )
+                    .padding(.leading, 40)
+
+                Spacer()
+
+                if viewModel.showAssetDetails {
+                    assetDetails
+                }
+            }
+            .padding(.bottom, 40)
         }
+        .transition(.opacity)
+    }
 
-        clearErrors?.cancel()
+    private var assetDetails: some View {
+        VStack(alignment: .trailing, spacing: 4) {
+            Text("\(viewModel.userAssetIndex) / \(viewModel.userAssetsCount)")
+                .font(.title2)
+                .foregroundColor(.white)
+                .shadow(color: .black.opacity(0.8), radius: 2, x: 0, y: 1)
 
-        let workItem = DispatchWorkItem {
-            withAnimation {
-                errors.removeAll()
+            if !viewModel.userDateTime.isEmpty {
+                Text(viewModel.userDateTime)
+                    .font(.headline)
+                    .foregroundColor(.white.opacity(0.9))
+                    .shadow(color: .black.opacity(0.7), radius: 1, x: 0, y: 1)
+            }
+
+            if !viewModel.userLocation.isEmpty {
+                Text(viewModel.userLocation)
+                    .font(.headline)
+                    .foregroundColor(.white.opacity(0.9))
+                    .shadow(color: .black.opacity(0.7), radius: 1, x: 0, y: 1)
             }
         }
-        clearErrors = workItem
-        DispatchQueue.main.asyncAfter(deadline: .now() + 15, execute: workItem)
-    }
-
-    func showInformation(_ message: String) {
-        AppLog.shared.log(message, level: .info, source: "SlideshowView")
-        let easterEggs = [
-            message,
-            "=^..^=",
-            "‘-‘_@_",
-            "----{,_,\">",
-            "><((((\">",
-            "~~(^._.^)~~",
-            ",/\\,/\\,/\\,/\\,/\\,/\\,o",
-            "<(‘–‘)>",
-            "///\\\\oo/\\\\\\"
-        ]
-        for (i, egg) in easterEggs.enumerated() {
-            let count = informations.count { $0 == egg }
-
-            if count >= 5 {
+        .padding(16)
+        .background(Color.black.opacity(0.6))
+        .cornerRadius(10)
+        .onAppear {
+            assetDetailsTimer?.invalidate()
+            assetDetailsTimer = Timer.scheduledTimer(
+                withTimeInterval: 10,
+                repeats: false
+            ) { _ in
                 withAnimation {
-                    informations.removeAll { $0 == egg }
-                    if easterEggs.indices.contains(i + 1) {
-                        informations.append(easterEggs[i + 1])
-                    }
+                    viewModel.showAssetDetails = false
                 }
             }
         }
-
-        withAnimation {
-            informations.append(message)
+        .onDisappear {
+            assetDetailsTimer?.invalidate()
+            assetDetailsTimer = nil
         }
     }
 
-    private func handleMoveCommand(_ direction: MoveCommandDirection) async {
-        switch direction {
-        case .left:
-            stopSlideshowTimer()
-            stopProgressBarTimer()
-
-            switch settings.slideshowLeftAction {
-            case .goToNext:
-                await moveToNext()
-            case .goToPrevious:
-                await moveToPrevious()
-            }
-        case .right:
-            stopSlideshowTimer()
-            stopProgressBarTimer()
-
-            switch settings.slideshowRightAction {
-            case .goToNext:
-                await moveToNext()
-            case .goToPrevious:
-                await moveToPrevious()
-            }
-        case .up:
-            if let player = currentPlayer {
-                seek(player: player, seconds: 15)
-            }
-        case .down:
-            if let player = currentPlayer {
-                seek(player: player, seconds: -15)
-            }
-        default:
-            break
-        }
-    }
-
-    private func seek(player: AVPlayer, seconds: Double) {
-        guard let currentItem = player.currentItem else { return }
-        let currentTime = player.currentTime()
-        let newTime = CMTimeAdd(
-            currentTime,
-            CMTimeMakeWithSeconds(
-                seconds,
-                preferredTimescale: currentTime.timescale
-            )
-        )
-
-        let clampedTime: CMTime
-        if CMTimeCompare(newTime, .zero) < 0 {
-            clampedTime = .zero
-        } else if CMTimeCompare(newTime, currentItem.duration) > 0 {
-            return
-        } else {
-            clampedTime = newTime
-        }
-
-        player.seek(to: clampedTime)
-    }
-
-    private func togglePause() {
-        if let player = currentPlayer {
-            let isPlaying =
-                player.timeControlStatus == .playing && player.rate != 0
-
-            if isPlaying {
-                player.pause()
-                videoPausedByUser = true
-            } else {
-                player.play()
-                videoPausedByUser = false
-                observeVideoProgress()
-            }
-        } else if currentImage != nil {
-            withAnimation {
-                slideshowIsRunning.toggle()
-            }
-
-            if slideshowIsRunning {
-                startImageTimers()
-            } else {
-                showAssetDetails = true
-                stopSlideshowTimer()
-                stopProgressBarTimer()
-            }
-        }
-    }
-
-    private func loadCurrentAsset(slideshowAsset: SlideshowAsset) async {
-        // stop actions
-        stopSlideshowTimer()
-        stopProgressBarTimer()
-        stopCurrentPlayer()
-
-        let asset = slideshowAsset.asset
-        self.slideshowAsset = slideshowAsset
-
-        if previousAlbumID == nil {
-            previousAlbumID = slideshowAsset.album.id
-        }
-        withAnimation {
-            informations.removeAll()
-        }
-
-        // clear state
-        currentImage = nil
-        currentPlayer = nil
-        assetProgress = 0.0
-
-        userAssetIndex = slideshowAsset.counter.current
-        userAssetsCount = slideshowAsset.counter.total
-        userDateTime = formattedCaptureDate(asset)
-        userLocation = formattedLocation(asset)
-
-        do {
-            if asset.assetType == .image {
-                if let cache = imageCache, let nextImage = cache.get(asset.id) {
-                    currentImage = nextImage
-                } else {
-                    let data = try await ImmichAPI.shared.loadMediaWithRetries(
-                        path:
-                        "/api/assets/\(slideshowAsset.asset.id.string)/thumbnail",
-                        queryParams: ["size": "fullsize"],
-                        retries: 3
-                    )
-                    if let uiImage = UIImage(data: data) {
-                        currentImage = Image(uiImage: uiImage)
-                    } else {
-                        showError("loading image failed: id=\(asset.id)")
-                        await moveToNext()
-                        return
-                    }
-                }
-
-                if slideshowIsRunning {
-                    startImageTimers()
-                }
-            } else if asset.assetType == .video {
-                var playbackURL: URL
-                do {
-                    playbackURL = try await ImmichAPI.shared
-                        .getUrlWithQueryAuth(
-                            path:
-                            "/api/assets/\(asset.id.string)/video/playback",
-                            queryParams: nil
-                        )
-                } catch {
-                    showError(
-                        "loading video failed: failed to construct playback URL"
-                    )
-                    return
-                }
-
-                let playerItem = AVPlayerItem(url: playbackURL)
-                let player = AVPlayer(playerItem: playerItem)
-                currentPlayer = player
-                videoPausedByUser = false
-
-                let oldAssetID = asset.id
-                DispatchQueue.main.asyncAfter(deadline: .now() + 5.0) {
-                    [oldAssetID] in
-                    guard let player = currentPlayer else { return }
-                    guard let sAsset = self.slideshowAsset,
-                          sAsset.asset.id == oldAssetID
-                    else { return }
-
-                    if player.status == .failed
-                        || player.currentItem?.status == .failed
-                    {
-                        showError("video failed to load")
-                        Task {
-                            await moveToNext()
-                        }
-                        return
-                    } else if player.timeControlStatus != .playing, !videoPausedByUser {
-                        showError("video did not start playing")
-                        Task {
-                            await moveToNext()
-                        }
-                        return
-                    }
-                }
-
-                observeVideoProgress()
-
-                if let currentPlayer {
-                    playerObserver = NotificationCenter.default.addObserver(
-                        forName: .AVPlayerItemDidPlayToEndTime,
-                        object: currentPlayer.currentItem,
-                        queue: .main
-                    ) { _ in
-                        Task {
-                            await moveToNext()
-                        }
-                    }
-                }
-            }
-        } catch {
-            showError(error.localizedDescription)
-            logError(error)
-            isLoading = false
-            await moveToNext()
-            return
-        }
-
-        if previousAlbumID != slideshowAsset.album.id {
-            previousAlbumID = slideshowAsset.album.id
-            showInformation(
-                "switched to a new album: \(slideshowAsset.album.albumName.string)"
-            )
-        }
-
-        await preloadAssets()
-    }
-
-    private func preloadAssets() async {
-        guard let slideshow else { return }
-
-        if let laterAsset = try? await slideshow.previewNext() {
-            await preloadAsset(
-                asset: laterAsset.asset
-            )
-        }
-        if let earlierAsset = try? await slideshow.previewPrevious() {
-            await preloadAsset(
-                asset: earlierAsset.asset
-            )
-        }
-    }
-
-    private func preloadAsset(asset: AlbumAsset) async {
-        if let cache = imageCache, asset.assetType == .image {
-            guard cache.get(asset.id) == nil else { return }
-
-            do {
-                let data = try await ImmichAPI.shared.loadMediaWithRetries(
-                    path: "/api/assets/\(asset.id.string)/thumbnail",
-                    queryParams: ["size": "fullsize"],
-                    retries: 2
+    private var progressBar: some View {
+        GeometryReader { geometry in
+            Rectangle()
+                .fill(Color.gray.opacity(0.6))
+                .frame(
+                    width: geometry.size.width * viewModel.assetProgress,
+                    height: 4
                 )
-                if let uiImage = UIImage(data: data) {
-                    cache.set(asset.id, value: Image(uiImage: uiImage))
+                .position(
+                    x: geometry.size.width / 2,
+                    y: geometry.size.height - 2
+                )
+        }
+        .ignoresSafeArea()
+    }
+
+    private var messagesOverlay: some View {
+        VStack {
+            Spacer()
+            HStack {
+                Spacer()
+                VStack(alignment: .trailing, spacing: 6) {
+                    ForEach(viewModel.errors.indices, id: \.self) { index in
+                        Text(viewModel.errors[index])
+                            .foregroundColor(.white)
+                            .padding(.horizontal, 12)
+                            .padding(.vertical, 8)
+                            .background(Color.red.opacity(0.8))
+                            .cornerRadius(10)
+                            .shadow(radius: 3)
+                    }
+
+                    ForEach(viewModel.informations.indices, id: \.self) { index in
+                        Text(viewModel.informations[index])
+                            .foregroundColor(.white)
+                            .padding(.horizontal, 12)
+                            .padding(.vertical, 8)
+                            .background(Color.blue.opacity(0.8))
+                            .cornerRadius(10)
+                            .shadow(radius: 3)
+                    }
                 }
-            } catch {
-                logError(error)
-                return
+                .padding(.trailing, 20)
+                .padding(.bottom, 20)
             }
         }
-
-        // TODO: add preloading videos, removed since it was too flaky
-        // it must be run in a background thread, it was messing up progress bar otherwise
-    }
-
-    private func startImageTimers() {
-        stopSlideshowTimer()
-        stopProgressBarTimer()
-        assetProgress = 0.0
-
-        let step = 0.05
-        let totalSteps = Double(settings.slideshowInterval) / step
-        progressBarTimer = Timer.scheduledTimer(
-            withTimeInterval: step,
-            repeats: true
-        ) { t in
-            assetProgress += 1 / totalSteps
-            if assetProgress >= 1 {
-                t.invalidate()
-            }
-        }
-
-        slideshowTimer = Timer.scheduledTimer(
-            withTimeInterval: TimeInterval(settings.slideshowInterval),
-            repeats: false
-        ) { _ in
-            Task {
-                await moveToNext()
-            }
-        }
-    }
-
-    private func stopSlideshowTimer() {
-        slideshowTimer?.invalidate()
-        slideshowTimer = nil
-    }
-
-    private func stopProgressBarTimer() {
-        progressBarTimer?.invalidate()
-        progressBarTimer = nil
-    }
-
-    private func observeVideoProgress() {
-        stopProgressBarTimer()
-        removeTimeObserver()
-        guard let player = currentPlayer else { return }
-
-        let interval = CMTime(
-            seconds: 0.05,
-            preferredTimescale: CMTimeScale(NSEC_PER_SEC)
-        )
-        timeObserverToken = player.addPeriodicTimeObserver(forInterval: interval, queue: .main) {
-            time in
-            if let duration = player.currentItem?.duration.seconds, duration > 0 {
-                assetProgress = min(time.seconds / duration, 1.0)
-            }
-        }
-    }
-
-    private func removeTimeObserver() {
-        if let timeObserverToken {
-            currentPlayer?.removeTimeObserver(timeObserverToken)
-            self.timeObserverToken = nil
-        }
-    }
-
-    private func stopCurrentPlayer() {
-        removeTimeObserver()
-        if let playerObserver {
-            NotificationCenter.default.removeObserver(playerObserver)
-            self.playerObserver = nil
-        }
-        currentPlayer?.pause()
-        currentPlayer = nil
-    }
-
-    private func moveToNext() async {
-        guard let slideshow else { return }
-
-        var next: SlideshowAsset?
-        do {
-            next = try await slideshow.next()
-        } catch {
-            showError("couldn't fetch the next asset: \(error)")
-            return
-        }
-
-        if let next {
-            stopCurrentPlayer()
-
-            Task {
-                await loadCurrentAsset(slideshowAsset: next)
-            }
-        } else {
-            // later asset doesn't exist
-            stopSlideshowTimer()
-            stopProgressBarTimer()
-            currentPlayer?.pause()
-            assetProgress = 0
-
-            if settings.slideshowDirection == .newestToOldest {
-                showInformation(
-                    "can't go back, this is the first image / video"
-                )
-            } else {
-                showInformation("the end!")
-            }
-        }
-    }
-
-    private func moveToPrevious() async {
-        guard let slideshow else { return }
-
-        var previous: SlideshowAsset?
-        do {
-            previous = try await slideshow.previous()
-        } catch {
-            showError("couldn't fetch the previous asset: \(error)")
-            return
-        }
-
-        if let previous {
-            stopCurrentPlayer()
-
-            Task {
-                await loadCurrentAsset(slideshowAsset: previous)
-            }
-        } else {
-            // earlier asset doesn't exist
-            stopSlideshowTimer()
-            stopProgressBarTimer()
-            currentPlayer?.pause()
-            assetProgress = 0
-
-            if settings.slideshowDirection == .newestToOldest {
-                showInformation("the end!")
-            } else {
-                showInformation(
-                    "can't go back, this is the first image / video"
-                )
-            }
-        }
-    }
-
-    private func initSlideshow(albumID _: AlbumID) async {
-        do {
-            slideshow = try await SlideshowSequencer(
-                playlistGetter: SlideshowPlaylistGetter(
-                    settings: settings,
-                    immichClient: ImmichClient.shared
-                ),
-                initialAlbumID: initialAlbumID,
-                initialAssetID: initialAssetID
-            )
-            guard let slideshow else { return }
-
-            if let currentAsset = try await slideshow.current() {
-                await loadCurrentAsset(slideshowAsset: currentAsset)
-            }
-        } catch {
-            showError(error.localizedDescription)
-            logError(error)
-        }
+        .transition(.opacity.combined(with: .move(edge: .trailing)))
+        .animation(.easeInOut, value: viewModel.errors)
     }
 }
